Okay never mind, not using Code Prettify anymore. This works now.
Part 2:

The man looked around, knowing exactly what he's looking for.

First, kai leaves. Its special long and thin shape, along with the slight stickiness of the surface of the leaves makes it ideal to be shaped into a container.

Second, sap from zeyda trees. It provides a sweet aroma that children love. It also serves to further bind the other ingredients together.

Finally, saitao berries. It has a bitter and sour taste, but it's said to have  healing properties. If you're careful, that is. Otherwise it's also a very potent killer.

The old man carried these back to the cave and set them down on a rectangular piece of flat rock that was already placed above a bundle of sticks. He prepared and arranged the ingredients.

With a snap, the branches lit aflame.