Cleaned up the code a little
Part 3:

The old man took out a long stone out of a cloth bundle, and used it to crush the berries in the kai leaves bowl.

The berries have to be roasted for a very precise amount of time. Usually this is done only by trained apocatharies, but the old man knew that no apocathary would be found in this forest. It was either try something, or watch the girl die.

Soon, the old man took the bowl out from the stone plate and wrapped the berries and the sap up with the leaves.

The girl stirred in the pile of leaves, and with a bit of struggle, opened her eyes. She turned her head to look at the old man. He smiled. A look of fear grew on her face, and started to back away.

The man held out a hand to stop her, and held out the leaf roll. Cautiously, she crept back forward and even more cautiously, took the roll. She looked at the roll, and inspected it with curiosity.

Finally, she took a bite.