Ported the CSS over too
I'm not really sure why, but the CSS doesn't 100% work for some pages. I'll figure this out later.

Part 5:

It was late afternoon by the time they arrived at the city hall.

The stone building stood tall with five levels. Its exterior was covered with ivy, but you could still see the sign in the front that proudly displayed the city's name. It was commissioned by the mayor 50 years ago when the town grew big enough that there needed a way for the government to convene. Over time, it also became a place for people to post and look for odd jobs.

The two walked through the large rectanglar door way. To their left, they saw the massive wooden board that had all the job postings.

"HORSE FOR SALE"
"HEAVY LIFTERS WANTED"
"LOOKING FOR A GOOD TIME? VISIT US!"
"LOOKING FOR LOST HORSE"

The old man scanned through the posters, looking for one in search of a lost girl.

Finally, the old man opted to write up his own post. He was told to check back in a few days, as that was how long a posting usually stayed up for before someone answers it.

But he wasn't too hopeful. If her parents really were looking, the two would be heading to their house right now.