Fixed CSS for real this time
It turns out that all the pages share CSS because of the way routing works. So that was one of the problems.

Part 7:

He spent his whole life alone, and expected it to stay that way forever.

The old man thought back to the time, long ago, when he was given a chance to change that life of solitude. But he was perfectly fine with being alone. And that's how he got into this mess in the first place.

But it'll all change from now on. He'll make the best of the situation. No more pasts, no more messes.

After all, a man like him rarely makes mistakes. Everything just requires careful thought, and who is better at careful thinking than an old man who's seen most of life?

Plus, raising a child can't be that hard, right?