More upload stuff
Uploading multiple files at once "works" now.

Part 9:

The old man and the girl approached the school. The building was a simple wooden rectangle. The school yard was decently sized, though it was also fairly simple. It was a bare lawn with a couple of trees here and there.

The pair stopped at the open door of the school. The old man let the girl go. Reluctantly, she went and he waved as he watched her go into her classroom.

Finally, the old man turned back into the streets and instead of heading home, he turned in the opposite direction.

He looked up at the cathedral looming over him. His eyes were fixed on the large stained glass window above the doors of the building. It depicted a woman with long flowing red hair. Otherwise, she looked like a generic woman. The intention is that every person looking at the window can feel like they know her personally. The idea is that all the love in the world is a reflection of her love.

The only difference with the old man is that he really did know her personally.